refactor: service layer 삭제 및 관련 로직 컨트롤러로 이동, 파일 이동
